AutoFaucet/requirements.txt
# 核心依赖
selenium>=4.15.0
undetected-chromedriver>=3.5.4
requests>=2.31.0
Pillow>=10.0.0
coloredlogs>=15.0.1
PyYAML>=6.0.1
fake-useragent>=1.4.0
python-dateutil>=2.8.2
cryptography>=41.0.0
PySocks>=1.7.1